backend/annotation-service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.32.3
Pillow==10.3.0
gunicorn==22.0.0
drf-spectacular==0.27.2
minio==7.2.9
